Add ability to run command before passing on TERM signal
This change adds the ability to run an arbitrary command between receiving a
TERM signal and passing it on to the child process(es).

This provides a mechanism to pull a node out of the load balancer or
similar, by returning an error on a health check URL before returning errors
on other endpoints.  Without this, if the load balancer doesn't have the
ability (or isn't configured) to retry failed requests on a healthy node,
errors may be passed through to end users until the load balancer notices
that it's offline.